+++ CHANGELOG.md
@@ -4,7 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+- [enhancement] Log file can be disabled #20
+
 ## 1.1.0 (2016-09-02)
+
 - [enhancement] - Removed Hashie as a dependency #12
 - [feature] Allow configuring consumers min_bytes & max_wait_time #15
 - [feature] Allow configuring producers max_queue_size, delivery_threshold & delivery_interval #16