+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## 1.2.0 (2016-10-10)
 
 - [enhancement] Log file can be disabled #20
 - [enhancement] Add a new extra (time_elapsed) for notifications "listener.process_message" and "listener.process_batch" #24
@@ -18,4 +18,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - [feature] Allow configuring force_encoding for message payload #18
 
 ## 1.0.0 (2016-08-08)
+
 - Published on Github with full fledged consumers and producers